requirements.txt
python-telegram-bot>=21.0,<23.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
selenium>=4.20.0
webdriver-manager>=4.0.0
Pillow>=10.0.0
urllib3>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
telethon>=1.35.0
paapi5-python-sdk @ https://github.com/hirotoshi/paapi5-python-sdk/archive/refs/heads/master.zip
